+++ src/RoadCaptain.App.RouteBuilder/DummyUserPreferences.cs
@@ -11,7 +11,6 @@ namespace RoadCaptain.App.RouteBuilder
         public string? Route { get; set; }
         public Point? InGameWindowLocation { get; set; } = new Point(600, 200);
         public bool EndActivityAtEndOfRoute { get; set; }
-        public bool LoopRouteAtEndOfRoute { get; set; }
         public Version LastOpenedVersion { get; set; } = new Version(0, 0, 0, 0);
 
         public byte[]? ConnectionSecret { get; }