+++ src/RoadCaptain/IUserPreferences.cs
@@ -10,7 +10,6 @@ namespace RoadCaptain
         public string? Route { get; set; }
         Point? InGameWindowLocation { get; set; }
         public bool EndActivityAtEndOfRoute { get; set; }
-        public bool LoopRouteAtEndOfRoute { get; set; }
         Version LastOpenedVersion { get; set; }
         public byte[]? ConnectionSecret { get; }
         public void Load();